+++ src/test/java/com/cloudogu/scmmanager/ScmMigratedV1NotifierTest.java
@@ -3,7 +3,7 @@ package com.cloudogu.scmmanager;
 import com.cloudogu.scmmanager.info.JobInformation;
 import hudson.model.Run;
 import okhttp3.OkHttpClient;
-import okhttp3.mockwebserver.MockResponse;
+import okhttp3.mockwebserver.Dispatcher;
 import okhttp3.mockwebserver.MockWebServer;
 import org.junit.Before;
 import org.junit.Test;
@@ -20,6 +20,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
 
 import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -38,6 +39,13 @@ public class ScmMigratedV1NotifierTest {
   private ScmV2NotifierProvider v2NotifierProvider;
 
   @Before
+  public void setUpServerAndClient() throws IOException {
+    Dispatcher mDispatcher = new RecordedRequestDispatcher();
+    server.setDispatcher(mDispatcher);
+    server.start();
+  }
+
+  @Before
   public void prepareAuthentication() {
     when(authenticationFactory.createHttp(run, "one"))
       .thenReturn(response -> response.header("Auth", "Awesome"));
@@ -45,14 +53,12 @@ public class ScmMigratedV1NotifierTest {
 
   @Test
   public void testNotifyWithoutMatchingV2Location() throws InterruptedException, IOException {
-    stubResource();
-
     CountDownLatch cdl = new CountDownLatch(1);
 
     ScmMigratedV1Notifier notifier = createV1Notifier();
     AtomicReference<JobInformation> reference = applyV2Notifier(cdl, null);
 
-    OkHttpClient client = new OkHttpClient();
+    OkHttpClient client = new OkHttpClient.Builder().followRedirects(false).build();
     notifier.setClient(client);
     notifier.notify("abc123", BuildStatus.success("old-repo", "Old-Repo", "https://oss.cloudogu.com"));
 
@@ -63,8 +69,6 @@ public class ScmMigratedV1NotifierTest {
 
   @Test
   public void testNotify() throws InterruptedException, IOException {
-    stubResource();
-
     CountDownLatch cdl = new CountDownLatch(1);
 
     ScmMigratedV1Notifier notifier = createV1Notifier();
@@ -74,7 +78,7 @@ public class ScmMigratedV1NotifierTest {
 
     BuildStatus success = BuildStatus.success("old-repo", "Old-Repo",  "https://oss.cloudogu.com");
 
-    OkHttpClient client = new OkHttpClient();
+    OkHttpClient client = new OkHttpClient.Builder().followRedirects(false).build();
     notifier.setClient(client);
     notifier.notify("abc123", success);
 
@@ -84,19 +88,11 @@ public class ScmMigratedV1NotifierTest {
     Mockito.verify(v2Notifier).notify("abc123", success);
   }
 
-  private void stubResource() throws IOException {
-    server.enqueue(
-      new MockResponse()
-        .addHeader("Location", "https://scm.scm-manager.org/scm/old/repo")
-        .setResponseCode(301)
-    );
-    server.start();
-  }
-
   private void assertInfo(AtomicReference<JobInformation> reference) {
     JobInformation received = reference.get();
+    assertNotNull(received);
     assertEquals("git", received.getType());
-    assertEquals("https://scm.scm-manager.org/scm/old/repo", received.getUrl());
+    assertEquals("https://localhost/scm/old/repo", received.getUrl());
     assertEquals("abc123", received.getRevision());
     assertEquals("one", received.getCredentialsId());
   }
