+++ src/test/java/com/cloudogu/scmmanager/scm/api/ApiClientTestBase.java
@@ -1,20 +1,13 @@
 package com.cloudogu.scmmanager.scm.api;
 
 import com.cloudogu.scmmanager.HttpAuthentication;
-import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
+import com.cloudogu.scmmanager.RecordedRequestDispatcher;
 import okhttp3.OkHttpClient;
 import okhttp3.mockwebserver.Dispatcher;
-import okhttp3.mockwebserver.MockResponse;
 import okhttp3.mockwebserver.MockWebServer;
-import okhttp3.mockwebserver.RecordedRequest;
 import org.junit.Before;
 
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.FileReader;
 import java.io.IOException;
-import java.util.Arrays;
 
 public class ApiClientTestBase {
 
@@ -58,50 +51,4 @@ public class ApiClientTestBase {
     return pathInjection[pathInjectionIndex++];
   }
 
-  /**
-   * This uses json mapping files from the resources (src/test/resources/mappings) and
-   * answers requests with responses stored there.
-   * These files had been created using wiremock on the first run.
-   */
-  private static class RecordedRequestDispatcher extends Dispatcher {
-    private final File mappings = new File(getClass().getClassLoader().getResource("mappings").getFile());
-
-    @Override
-    public MockResponse dispatch(RecordedRequest request) {
-      String requestedPath = request.getPath();
-      return Arrays.stream(mappings.listFiles())
-        .filter(
-          file -> {
-            try {
-              JsonObject mapping = JsonParser.parseReader(new FileReader(file)).getAsJsonObject();
-              JsonObject recoredRequest = mapping.get("request").getAsJsonObject();
-              return requestedPath.endsWith(recoredRequest.get("url").getAsString())
-                && recoredRequest.get("headers").getAsJsonObject().get("Accept").getAsJsonObject().get("equalTo").getAsString().equals(request.getHeader("Accept"));
-            } catch (FileNotFoundException e) {
-              throw new RuntimeException("failed to read mapping " + file, e);
-            }
-          }
-        )
-        .map(
-          file -> {
-            try {
-              JsonObject mapping = JsonParser.parseReader(new FileReader(file)).getAsJsonObject();
-              JsonObject recordedResponse = mapping.get("response").getAsJsonObject();
-              MockResponse mockResponse = new MockResponse()
-                .setResponseCode(recordedResponse.get("status").getAsInt())
-                .setBody(recordedResponse.get("body").getAsString());
-              JsonObject headers = recordedResponse.get("headers").getAsJsonObject();
-              headers.keySet().forEach(
-                h -> mockResponse.setHeader(h, headers.get(h).getAsString())
-              );
-              return mockResponse;
-            } catch (FileNotFoundException e) {
-              throw new RuntimeException("failed to read mapping " + file, e);
-            }
-          }
-        )
-        .findFirst()
-        .orElseGet(() -> new MockResponse().setResponseCode(404));
-    }
-  }
 }
